pyproject.toml
[tool.poetry]
name = "sample-django-api"
version = "0.1.0"
description = "Starter Django project for backend developer test task – extend the API based on provided requirements."
authors = ["Artiom Cozlovici <kozakovi4@outlook.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"
gunicorn = "^23.0.0"
psycopg = {extras = ["binary"], version = "^3.2.9"}
environs = "^14.2.0"
factory-boy = "^3.3.3"
djangorestframework-simplejwt = "^5.5.0"
drf-spectacular = "^0.28.0"
drf-standardized-errors = "^0.15.0"
django-unfold = "^0.60.0"
django-filter = "^25.1"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
